Incorrect Param Cast
Integer parameters in d.net are 64 bit, not 32 bit. Also removed specialized cast exception catch that was meant to solve a problem that no longer exists.

+++ Infra/Commands.cs
@@ -43,7 +43,7 @@ namespace GraviOS.Infra
 		{
 			using (Wiki_DbContext ctxt = new())
 			{
-				int? taskId = (int?)arg.Data.Options?.FirstOrDefault()?.Value;
+				long? taskId = (long?)arg.Data.Options?.FirstOrDefault()?.Value;
 				WikiTask? task = ctxt.WikiTasks
 					.Include(x => x.Creator)
 					.Include(x => x.Updates)
